long_to_unsigned_int

+++ C/polynome.h
@@ -4,12 +4,12 @@
 #define NB_P 2013265921
 
 typedef struct _Poly {
-    long *coeffs;
+    unsigned int *coeffs;
     int deg;
 } Poly;
 
 void afficher_poly(Poly P);
-Poly gen_poly(int deg);
+Poly gen_poly(unsigned int deg);
 Poly liberer_poly(Poly P);
 int compare_poly(Poly P, Poly Q);
 Poly prod_poly_naif(Poly P, Poly Q);
@@ -17,13 +17,13 @@ Poly copy_poly(Poly P, int deb, int fin);
 Poly poly_somme(Poly P, Poly Q, int deb);
 Poly oppose_poly(Poly P);
 Poly prod_poly_karatsuba(Poly P, Poly Q);
-long mod_pow(long x, long n);
-long horner(Poly P, long x);
-long *get_racines(long racine, int n);
-long *eval(Poly P, long *racines);
-long mod_add(long a, long b, long p);
-long mod_sub(long a, long b, long p);
-long mod_mult(long a, long b, long p);
-long inv(long a, long p);
+unsigned int mod_pow(unsigned int x, unsigned int n);
+unsigned int horner(Poly P, unsigned int x);
+unsigned int *get_racines(unsigned int racine, int n);
+unsigned int *eval(Poly P, unsigned int *racines);
+unsigned int mod_add(unsigned int a, unsigned int b, unsigned int p);
+unsigned int mod_sub(unsigned int a, unsigned int b, unsigned int p);
+unsigned int mod_mult(unsigned int a, unsigned int b, unsigned int p);
+unsigned int inv(unsigned int a, unsigned int p);
 
 #endif
